refactor: move ValidatorOptions into its own file under CloudHealthOffice.Tools.MccPlatformValidator namespace
