+++ src/tools/mcc-platform-validator/ValidatorOptions.cs
@@ -1,0 +1,151 @@
+namespace CloudHealthOffice.Tools.MccPlatformValidator;
+
+public sealed record ValidatorOptions(
+    int Claims,
+    int Seed,
+    string TenantId,
+    string ClaimsUrl,
+    string BenefitUrl,
+    string ProviderUrl,
+    bool SeedProviders,
+    bool SkipClaimUpdate,
+    int TimeoutSeconds,
+    int ProgressEvery,
+    int Parallelism,
+    int LineOfBusiness,
+    string? SummaryJsonPath,
+    bool NoPublishSummary,
+    int PublishClaimResultsLimit,
+    bool PriorAuthScenariosEnabled,
+    double PriorAuthScenarioRate,
+    bool ShowHelp)
+{
+    public const int DefaultMaxClaims = 10_000;
+    public const int MaxParallelism = 64;
+
+    public static ValidatorOptions Parse(string[] args)
+    {
+        var options = new MutableOptions();
+        for (var i = 0; i < args.Length; i++)
+        {
+            switch (args[i])
+            {
+                case "--claims" or "-n" when i + 1 < args.Length:
+                    options.Claims = int.Parse(args[++i]);
+                    break;
+                case "--seed" or "-s" when i + 1 < args.Length:
+                    options.Seed = int.Parse(args[++i]);
+                    break;
+                case "--tenant" when i + 1 < args.Length:
+                    options.TenantId = args[++i];
+                    break;
+                case "--claims-url" when i + 1 < args.Length:
+                    options.ClaimsUrl = args[++i].TrimEnd('/');
+                    break;
+                case "--benefit-url" when i + 1 < args.Length:
+                    options.BenefitUrl = args[++i].TrimEnd('/');
+                    break;
+                case "--provider-url" when i + 1 < args.Length:
+                    options.ProviderUrl = args[++i].TrimEnd('/');
+                    break;
+                case "--no-seed-providers":
+                    options.SeedProviders = false;
+                    break;
+                case "--skip-claim-update":
+                    options.SkipClaimUpdate = true;
+                    break;
+                case "--timeout" when i + 1 < args.Length:
+                    options.TimeoutSeconds = int.Parse(args[++i]);
+                    break;
+                case "--progress-every" when i + 1 < args.Length:
+                    options.ProgressEvery = int.Parse(args[++i]);
+                    break;
+                case "--parallelism" or "-p" when i + 1 < args.Length:
+                    options.Parallelism = int.Parse(args[++i]);
+                    break;
+                case "--max-claims" when i + 1 < args.Length:
+                    options.MaxClaims = int.Parse(args[++i]);
+                    break;
+                case "--line-of-business" when i + 1 < args.Length:
+                    options.LineOfBusiness = int.Parse(args[++i]);
+                    break;
+                case "--no-prior-auth-scenarios":
+                    options.PriorAuthScenariosEnabled = false;
+                    break;
+                case "--prior-auth-rate" when i + 1 < args.Length:
+                    options.PriorAuthScenarioRate = double.Parse(args[++i], System.Globalization.CultureInfo.InvariantCulture);
+                    break;
+                case "--summary-json" when i + 1 < args.Length:
+                    options.SummaryJsonPath = args[++i];
+                    break;
+                case "--no-publish-summary":
+                    options.NoPublishSummary = true;
+                    break;
+                case "--claim-results-limit" when i + 1 < args.Length:
+                    options.PublishClaimResultsLimit = int.Parse(args[++i]);
+                    break;
+                case "--help" or "-h":
+                    options.ShowHelp = true;
+                    break;
+            }
+        }
+
+        var maxClaims = Math.Max(1, options.MaxClaims);
+        if (options.Claims > maxClaims)
+        {
+            Console.Error.WriteLine($"warning: capping --claims to {maxClaims:N0} to avoid excessive in-memory allocation");
+            options.Claims = maxClaims;
+        }
+
+        if (options.Parallelism > MaxParallelism)
+        {
+            Console.Error.WriteLine($"warning: capping --parallelism to {MaxParallelism:N0} for local validation");
+            options.Parallelism = MaxParallelism;
+        }
+
+        var effectiveClaims = Math.Max(1, options.Claims);
+
+        return new ValidatorOptions(
+            effectiveClaims,
+            options.Seed,
+            options.TenantId,
+            options.ClaimsUrl.TrimEnd('/'),
+            options.BenefitUrl.TrimEnd('/'),
+            options.ProviderUrl.TrimEnd('/'),
+            options.SeedProviders,
+            options.SkipClaimUpdate,
+            Math.Max(5, options.TimeoutSeconds),
+            Math.Max(1, options.ProgressEvery),
+            Math.Max(1, options.Parallelism),
+            Math.Clamp(options.LineOfBusiness, 1, 5),
+            options.SummaryJsonPath,
+            options.NoPublishSummary,
+            Math.Clamp(options.PublishClaimResultsLimit, 0, effectiveClaims),
+            options.PriorAuthScenariosEnabled,
+            Math.Clamp(options.PriorAuthScenarioRate, 0.0, 0.25),
+            options.ShowHelp);
+    }
+
+    private sealed class MutableOptions
+    {
+        public int Claims { get; set; } = 25;
+        public int Seed { get; set; } = 42;
+        public string TenantId { get; set; } = "demo";
+        public string ClaimsUrl { get; set; } = "http://localhost:5001";
+        public string BenefitUrl { get; set; } = "http://localhost:5002";
+        public string ProviderUrl { get; set; } = "http://localhost:5004";
+        public bool SeedProviders { get; set; } = true;
+        public bool SkipClaimUpdate { get; set; }
+        public int TimeoutSeconds { get; set; } = 60;
+        public int ProgressEvery { get; set; } = 10;
+        public int Parallelism { get; set; } = 10;
+        public int MaxClaims { get; set; } = DefaultMaxClaims;
+        public int LineOfBusiness { get; set; } = 3;
+        public string? SummaryJsonPath { get; set; }
+        public bool NoPublishSummary { get; set; }
+        public int PublishClaimResultsLimit { get; set; } = 1000;
+        public bool PriorAuthScenariosEnabled { get; set; } = true;
+        public double PriorAuthScenarioRate { get; set; } = 0.02;
+        public bool ShowHelp { get; set; }
+    }
+}
